tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/keyframe_empty_block.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - keyframe rule with empty block</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="A keyframe with an empty block is valid">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 100px;
         }
         to {
            width: 200px;
         }
         to {
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation: grow 0.1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = (getComputedStyle(div, null).width == '100px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if no animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should be square and stay that way.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
